Refactor and reorganize package structure.
Moved `command` and `formater` packages into `pkg/core` for better modularity and alignment with project architecture. Updated imports and references accordingly to ensure functionality remains intact.
