+++ pkg/core/command/errors_test.go
