+++ pkg/core/command/factory.go
