+++ pkg/core/command/factory_test.go
