+++ pkg/core/command/macro.go
