+++ pkg/core/command/macro_test.go
