+++ pkg/core/formater/formater_test.go
