+++ pkg/core/formater/text.go
