+++ pkg/core/formater/text_test.go
